+++ src/main/java/cl/duoc/ferremasapp/dto/PagoDTO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class PagoDTO {
-    private Long pedidoId;
+    private Integer pedidoId;
     private double monto;
     private String metodoPago; // DEBITO, CREDITO, TRANSFERENCIA
 }
