+++ src/main/java/cl/duoc/ferremasapp/repository/CategoriaRepository.java
@@ -1,6 +1,7 @@
 package cl.duoc.ferremasapp.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
 import cl.duoc.ferremasapp.model.Categoria;
 
-public interface CategoriaRepository extends JpaRepository<Categoria, Long> {}
+public interface CategoriaRepository extends JpaRepository<Categoria, Integer> {}
