+++ src/main/java/cl/duoc/ferremasapp/repository/MarcaRepository.java
@@ -1,6 +1,7 @@
 package cl.duoc.ferremasapp.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
 import cl.duoc.ferremasapp.model.Marca;
 
-public interface MarcaRepository extends JpaRepository<Marca, Long> {}
+public interface MarcaRepository extends JpaRepository<Marca, Integer> {}
