+++ src/main/java/cl/duoc/ferremasapp/repository/PagoRepository.java
@@ -1,6 +1,7 @@
 package cl.duoc.ferremasapp.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
 import cl.duoc.ferremasapp.model.Pago;
 
-public interface PagoRepository extends JpaRepository<Pago, Long> {}
+public interface PagoRepository extends JpaRepository<Pago, Integer> {}
