+++ src/main/java/cl/duoc/ferremasapp/repository/PedidoRepository.java
@@ -1,6 +1,7 @@
 package cl.duoc.ferremasapp.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
 import cl.duoc.ferremasapp.model.Pedido;
 
-public interface PedidoRepository extends JpaRepository<Pedido, Long> {}
+public interface PedidoRepository extends JpaRepository<Pedido, Integer> {}
