+++ src/main/java/cl/duoc/ferremasapp/repository/RolRepository.java
@@ -1,6 +1,7 @@
 package cl.duoc.ferremasapp.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
 import cl.duoc.ferremasapp.model.Rol;
 
-public interface RolRepository extends JpaRepository<Rol, Long> {}
+public interface RolRepository extends JpaRepository<Rol, Integer> {}
