+++ src/main/java/cl/duoc/ferremasapp/repository/StockRepository.java
@@ -1,6 +1,7 @@
 package cl.duoc.ferremasapp.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
 import cl.duoc.ferremasapp.model.Stock;
 
-public interface StockRepository extends JpaRepository<Stock, Long> {}
+public interface StockRepository extends JpaRepository<Stock, Integer> {}
