+++ src/main/java/cl/duoc/ferremasapp/repository/SucursalRepository.java
@@ -1,6 +1,7 @@
 package cl.duoc.ferremasapp.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
 import cl.duoc.ferremasapp.model.Sucursal;
 
-public interface SucursalRepository extends JpaRepository<Sucursal, Long> {}
+public interface SucursalRepository extends JpaRepository<Sucursal, Integer> {}
